refactor components -> individual folders
